+++ core-ng/src/main/java/core/framework/module/APIConfig.java
@@ -2,7 +2,6 @@ package core.framework.module;
 
 import core.framework.api.web.service.Path;
 import core.framework.http.HTTPClient;
-import core.framework.http.HTTPClientBuilder;
 import core.framework.http.HTTPMethod;
 import core.framework.internal.inject.InjectValidator;
 import core.framework.internal.module.Config;
@@ -18,12 +17,10 @@ import core.framework.internal.web.service.WebServiceImplValidator;
 import core.framework.internal.web.service.WebServiceInterfaceValidator;
 import core.framework.util.ASCII;
 import core.framework.web.Controller;
-import core.framework.web.service.WebServiceClientInterceptor;
 import core.framework.web.service.WebServiceClientProxy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
 import java.lang.reflect.Method;
 import java.time.Duration;
 
@@ -33,7 +30,6 @@ import java.time.Duration;
 public class APIConfig extends Config {
     private final Logger logger = LoggerFactory.getLogger(APIConfig.class);
     ModuleContext context;
-    private HTTPClientBuilder httpClientBuilder;
     private HTTPClient httpClient;
     private RequestBeanWriter writer;
     private ResponseBeanReader reader;
@@ -42,14 +38,15 @@ public class APIConfig extends Config {
     protected void initialize(ModuleContext context, String name) {
         this.context = context;
         // default value is for internal api call only, targeting for kube env (with short connect timeout and more retries)
-        httpClientBuilder = HTTPClient.builder()
+        httpClient = HTTPClient.builder()
             .userAgent(WebServiceClient.USER_AGENT)
             .trustAll()
             .connectTimeout(Duration.ofSeconds(2))
             .timeout(Duration.ofSeconds(20))    // refer to: kube graceful shutdown period is 30s, db timeout is 15s
             .keepAlive(Duration.ofMinutes(5))   // use longer keep alive timeout within cluster, to reduce connection creation overhead
             .slowOperationThreshold(Duration.ofSeconds(10))
-            .maxRetries(5);
+            .maxRetries(5)
+            .build();
         writer = new RequestBeanWriter();
         reader = new ResponseBeanReader();
     }
@@ -81,12 +78,16 @@ public class APIConfig extends Config {
     }
 
     public <T> APIClientConfig client(Class<T> serviceInterface, String serviceURL) {
-        T client = createClient(serviceInterface, serviceURL, getOrCreateHTTPClient());
+        return client(serviceInterface, serviceURL, httpClient);
+    }
+
+    public <T> APIClientConfig client(Class<T> serviceInterface, String serviceURL, HTTPClient httpClient) {
+        T client = createClient(serviceInterface, serviceURL, httpClient);
         context.beanFactory.bind(serviceInterface, null, client);
         return new APIClientConfig((WebServiceClientProxy) client);
     }
 
-    private <T> T createClient(Class<T> serviceInterface, String serviceURL, HTTPClient httpClient) {
+    public <T> T createClient(Class<T> serviceInterface, String serviceURL, HTTPClient httpClient) {
         logger.info("create web service client, interface={}, serviceURL={}", serviceInterface.getCanonicalName(), serviceURL);
         var validator = new WebServiceInterfaceValidator(serviceInterface, context.beanClassValidator);
         validator.requestBeanWriter = writer;
@@ -95,27 +96,7 @@ public class APIConfig extends Config {
         return createWebServiceClient(serviceInterface, new WebServiceClient(serviceURL, httpClient, writer, reader));
     }
 
-    public <T> T createClient(Class<T> serviceInterface, String serviceURL, HTTPClient httpClient, @Nullable WebServiceClientInterceptor interceptor) {
-        T proxy = createClient(serviceInterface, serviceURL, httpClient);
-        if (interceptor != null) {
-            new APIClientConfig((WebServiceClientProxy) proxy).intercept(interceptor);
-        }
-        return proxy;
-    }
-
     <T> T createWebServiceClient(Class<T> serviceInterface, WebServiceClient webServiceClient) {
         return new WebServiceClientBuilder<>(serviceInterface, webServiceClient).build();
-    }
-
-    public HTTPClientBuilder httpClient() {
-        if (httpClient != null) throw new Error("api().httpClient() must be configured before adding client");
-        return httpClientBuilder;
-    }
-
-    private HTTPClient getOrCreateHTTPClient() {
-        if (httpClient == null) {
-            this.httpClient = httpClientBuilder.build();
-        }
-        return httpClient;
     }
 }